requirements.txt
absl-py==0.11.0
astor==0.8.1
backports.weakref==1.0.post1
click==7.1.2
enum34==1.1.10
Flask
funcsigs==1.0.2
futures==3.3.0
gast==0.4.0
google-pasta==0.2.0
grpcio==1.33.2
h5py==2.10.0
itsdangerous==1.1.0
Jinja2==2.11.2
Keras-Applications
Keras-Preprocessing
Markdown==3.1.1
MarkupSafe==1.1.1
mock==3.0.5
numpy==1.16.6
Pillow==6.2.2
protobuf==3.14.0
six==1.15.0
tensorboard==1.14.0
tensorflow
tensorflow-estimator==1.14.0
termcolor==1.1.0
Werkzeug==1.0.1
wrapt==1.12.1
opencv-python
waitress